test(#5025): retain evidence warnings in CI lane
