+++ src/services/discord/turn_bridge/terminal_outcome_delivery/delivery_epilogue.rs
@@ -60,15 +60,6 @@ pub(super) struct DeliveryEpilogueState<'a> {
     pub(super) busy_requeue_outcome: &'a mut Option<followup_requeue::FollowupRequeueOutcome>,
 }
 
-fn bridge_terminal_delivery_evidence_loss_should_warn(
-    outcome: crate::services::discord::inflight::GuardedSaveOutcome,
-) -> bool {
-    matches!(
-        outcome,
-        crate::services::discord::inflight::GuardedSaveOutcome::IdentityMismatch
-    )
-}
-
 #[rustfmt::skip]
 pub(super) async fn handle_delivery_epilogue(
     message: DeliveryEpilogueMessage,
@@ -125,31 +116,13 @@ pub(super) async fn handle_delivery_epilogue(
                     &mut *inflight_state,
                     "turn_bridge::terminal_delivery_committed_mirror@5536",
                 );
-            // #5025: the terminal answer WAS delivered, but the row that would
-            // record it may now be owned by another turn identity. Keep the
-            // identity gate's non-clobber behavior and make the lost mirror
-            // evidence observable instead of silently reporting "undelivered".
-            if bridge_terminal_delivery_evidence_loss_should_warn(mirror_outcome) {
-                tracing::warn!(
-                    provider = %provider.as_str(),
-                    channel_id = channel_id.get(),
-                    current_msg_id = current_msg_id.get(),
-                    response_sent_offset,
-                    "turn bridge delivered the terminal answer but could not mirror terminal_delivery_committed: the inflight row is owned by a different turn identity, so this turn's delivery evidence is lost (row will read as undelivered)"
-                );
-            }
-            match mirror_outcome {
-                crate::services::discord::inflight::GuardedSaveOutcome::IoError => {
-                    tracing::warn!(
-                        provider = %provider.as_str(),
-                        channel_id = channel_id.get(),
-                        "turn bridge failed to mirror committed terminal delivery before cleanup"
-                    );
-                }
-                crate::services::discord::inflight::GuardedSaveOutcome::Saved
-                | crate::services::discord::inflight::GuardedSaveOutcome::Missing
-                | crate::services::discord::inflight::GuardedSaveOutcome::IdentityMismatch => {}
-            }
+            super::warn_if_bridge_terminal_delivery_evidence_lost(
+                mirror_outcome,
+                &provider,
+                channel_id,
+                current_msg_id,
+                response_sent_offset,
+            );
             for frozen_msg_id in terminal_full_replay_cleanup_msg_ids.drain(..) {
                 // #5413/#3607: current_msg_id is the terminal answer and is
                 // already excluded here, so no terminal-anchor guard is
@@ -453,16 +426,4 @@ pub(super) async fn handle_delivery_epilogue(
     *state.status_panel_terminal_committed = status_panel_terminal_committed;
 
     DeliveryEpilogueOutcome::Continue
-}
-
-#[cfg(test)]
-mod terminal_delivery_evidence_loss_tests {
-    use super::*;
-
-    #[test]
-    fn identity_mismatch_bridge_mirror_warns_about_lost_delivery_evidence() {
-        assert!(bridge_terminal_delivery_evidence_loss_should_warn(
-            crate::services::discord::inflight::GuardedSaveOutcome::IdentityMismatch
-        ));
-    }
 }
